fixed test segregation
